01_wait_forever: Simplify: Remove raw assembly
This patch removes the need for start.S completely leaving a purer Rust
story in place.

To do this, the patch introduces start.rs which mimics the functionality
of start.S.

A function called start within start.S is decorated with:

'#[link_section = ".text.start"]'

.. which instructs the linker to create a .text.start section for this
function accordingly.

The linker script is modded to:

- Ensure that the entry point is the start fn.
- Ensure that code in the '.text.start' section is placed first in the
  .text section.

.. thus enabling the use of pure Rust.

The cortex-a crate is pulled in for use of the asm module within.

Cargo.lock and a pre-built kernel binary are included.